requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
playwright==1.54.0

# CrewAI core
crewai
crewai-tools

# LangChain + adapter cho Gemini
langchain>=0.2.14
langchain-google-genai>=0.1.6

# Nếu muốn dùng OpenAI song song
openai>=1.37.0
langchain-openai>=0.1.16
